Fixed buggy usage of downloadFile

+++ raspiBackup.sh
@@ -3072,8 +3072,9 @@ function downloadPropertiesFile() { # FORCE
 			local downloadURL="$PROPERTIES_DOWNLOAD_URL"
 		fi
 
-		local dlHttpCode=$(downloadFile "$downloadURL" "$LATEST_TEMP_PROPERTY_FILE")
-		local dlRC=$?
+		local dlHttpCode dlRC
+		dlHttpCode=$(downloadFile "$downloadURL" "$LATEST_TEMP_PROPERTY_FILE")
+		dlRC=$?
 		if (( $dlRC != 0 )); then
 			if [[ $1 == "FORCE" ]]; then
 				writeToConsole $MSG_LEVEL_MINIMAL $MSG_DOWNLOAD_FAILED "$(sed "s/\?.*$//" <<< "$downloadURL")" "$dlHttpCode" $dlRC
@@ -3188,6 +3189,7 @@ function verifyIsOnOff() { # arg
 }
 
 function downloadFile() { # url, targetFileName
+
 		logEntry "URL: "$(sed -E "s/\?.*$//" <<<"$1")", file: $2"
 		local url="$1"
 		local file="$2"
@@ -3196,20 +3198,30 @@ function downloadFile() { # url, targetFileName
 		local rc=$?
 		logItem "httpCode: $httpCode RC: $rc"
 
-		if [[ $rc != 0 || ${httpCode:0:1} != "2" ]]; then
-			(( $rc == 0 )) && rc=100
-		elif head -n 1 "$f" | grep -q "^<!DOCTYPE html>"; then						# Download plugin doesn't return 404 if file not found but a HTML doc
-			httpCode="404"
-			rc=101
+		# Some nasty code required because download plugin doesn't return 404 if file not found but a HTML doc
+
+		if (( $rc == 0 )); then
+			if [[ ! -f "$f" ]]; then
+					httpCode="404"
+					rc=101
+			elif [[ ${httpCode:0:1} == "2" ]]; then
+				if head -n 1 "$f" | grep -q "^<!DOCTYPE html>"; then
+					httpCode="404"
+					rc=101
+				fi
+			else
+				rc=101
+			fi
 		fi
+
 		if (( $rc != 0 )); then
-			rm $f &>>$LOG_FILE
+			[[ -f $f ]] && rm $f &>>$LOG_FILE
 			echo "$httpCode"
 			logExit "$rc $httpCode"
 			return $rc
 		fi
 
-		mv $f $file &>>$LOG_FILE
+		[[ -f $f ]] && mv $f $file &>>$LOG_FILE
 		echo "200"
 		logExit 0
 		return 0
@@ -3483,14 +3495,14 @@ function updateScript() {
 			local file="${MYSELF}"
 			writeToConsole $MSG_LEVEL_MINIMAL $MSG_DOWNLOADING "$file" "$MYHOMEURL"
 
-			local dlHttpCode="$(downloadFile "$DOWNLOAD_URL" "${MYSELF}~")"
-			local dlRC=$?
+			local dlHttpCode dlRC
+			dlHttpCode="$(downloadFile "$DOWNLOAD_URL" "${MYSELF}~")"
+			dlRC=$?
 			if (( $dlRC != 0 )); then
 				writeToConsole $MSG_LEVEL_MINIMAL $MSG_DOWNLOAD_FAILED "$$DOWNLOAD_URL" "$dlHttpCode" $dlRC
 				writeToConsole $MSG_LEVEL_MINIMAL $MSG_SCRIPT_UPDATE_FAILED "$MYSELF"
 				exitError $RC_DOWNLOAD_FAILED
 			fi
-
 			newName="$SCRIPT_DIR/$MYNAME.$oldVersion.sh"
 			mv $SCRIPT_DIR/$MYSELF $newName
 			mv $MYSELF~ $SCRIPT_DIR/$MYSELF
